updateLightness function for Slideform class
